Change int64 to int

+++ store.go
@@ -10,13 +10,13 @@ type Store interface {
 	CreateStorage() error
 
 	GetRegions(userId int) ([]*Region, error)
-	GetRegion(userId int, regionId int64) (Region, error)
+	GetRegion(userId int, regionId int) (Region, error)
 	GetCategorys(userId int) ([]*Category, error)
-	GetCategory(userId int, categoryId int64) (Category, error)
+	GetCategory(userId int, categoryId int) (Category, error)
 	GetExpenses(userId int) ([]*Expense, error)
-	GetExpense(userId int, expenseId int64) (Expense, error)
+	GetExpense(userId int, expenseId int) (Expense, error)
 	GetRecipients(userId int) ([]*Recipient, error)
-	GetRecipient(userId int, recipientId int64) (Recipient, error)
+	GetRecipient(userId int, recipientId int) (Recipient, error)
 
 	CreateRegion(region *Region) error
 	CreateCategory(category *Category) error
@@ -58,7 +58,7 @@ func (store *dbStore) GetRegions(userId int) ([]*Region, error) {
 	return regions, nil
 }
 
-func (store *dbStore) GetRegion(userId int, regionId int64) (Region, error) {
+func (store *dbStore) GetRegion(userId int, regionId int) (Region, error) {
 	region := Region{}
 	err := store.db.QueryRow("SELECT id, description FROM region WHERE user_id = ? AND id = ?", userId, regionId).Scan(&region.Id, &region.Description)
 	
@@ -86,7 +86,7 @@ func (store *dbStore) GetCategorys(userId int) ([]*Category, error) {
 	return categorys, nil
 }
 
-func (store *dbStore) GetCategory(userId int, categoryId int64) (Category, error) {
+func (store *dbStore) GetCategory(userId int, categoryId int) (Category, error) {
 	category := Category{}
 	err := store.db.QueryRow("SELECT id, description FROM category WHERE user_id = ? AND id = ?", userId, categoryId).Scan(&category.Id, &category.Description)
 	
@@ -114,7 +114,7 @@ func (store *dbStore) GetRecipients(userId int) ([]*Recipient, error) {
 	return recipients, nil
 }
 
-func (store *dbStore) GetRecipient(userId int, recipientId int64) (Recipient, error) {
+func (store *dbStore) GetRecipient(userId int, recipientId int) (Recipient, error) {
 	recipient := Recipient{}
 	err := store.db.QueryRow("SELECT id, name FROM recipient WHERE user_id = ? AND id = ?", userId, recipientId).Scan(&recipient.Id, &recipient.Name)
 	
@@ -142,7 +142,7 @@ func (store *dbStore) GetExpenses(userId int) ([]*Expense, error) {
 	return expenses, nil
 }
 
-func (store *dbStore) GetExpense(userId int, expenseId int64) (Expense, error) {
+func (store *dbStore) GetExpense(userId int, expenseId int) (Expense, error) {
 	expense := Expense{}
 	err := store.db.QueryRow("SELECT description, amount, date, category_id, region_id, recipient_id FROM expense WHERE user_id = ? AND id = ?", userId, expenseId).Scan(&expense.Description, &expense.Amount, &expense.Date, &expense.CategoryId, &expense.RegionId, &expense.RecipientId)
